Hoist the duplicated datadir fixture into conftest.py
The `datadir` fixture (copy a test module's data folder into a tmpdir) was
copy-pasted verbatim across four test modules (test_010, test_011, test_016,
test_019). Move it to conftest.py so all tests share one definition, and drop
the now-unused per-module imports (os, distutils.dir_util, pytest.fixture).

Also remove test_031's commented-out test_cc3_h2 (dead code in a triple-quoted
string -- the live test_cc3_h2o already uses the rhf_wfn fixture).

No behavior change: test_016/test_019 keep loading their pinned reference wfns
via Wavefunction.from_file (reproducibility-pinned checkpoint/local-CC/RT
tests), they just get `datadir` from conftest now.
