+++ pycc/tests/test_010_denoise.py
@@ -1,28 +1,7 @@
 # Import required packages
 import numpy as np
 import pytest
-import os
 from pycc.rt.utils import denoise
-from pytest import fixture
-from distutils import dir_util
-
-@fixture
-def datadir(tmpdir, request):
-    '''
-    from: https://stackoverflow.com/a/29631801
-    Fixture responsible for searching a folder with the same name of test
-    module and, if available, moving all contents to a temporary directory so
-    tests can use them freely.
-    '''
-    filename = request.module.__file__
-    test_dir, _ = os.path.splitext(filename)
-
-    if os.path.isdir(test_dir):
-        dir_util.copy_tree(test_dir, str(tmpdir))
-    else:
-        raise FileNotFoundError("Test folder not found.")
-
-    return tmpdir
 
 def test_denoise(datadir):
     # Define function
